tools/swagger-loadgen/pyproject.toml
[project]
name = "swagger-loadgen"
version = "0.1.0"
description = "Load generator that parses OpenAPI/Swagger specs and fires GET requests at a fixed TPS"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1",
    "typer>=0.15,<1",
    "pyyaml>=6,<7",
    "rich>=13,<14",
]

[project.scripts]
swagger-loadgen = "swagger_loadgen.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/swagger_loadgen"]

[tool.ruff]
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
